skills/project-knowledge-base/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anthropic==0.84.0
    # via -r requirements.in
anyio==4.12.1
    # via
    #   anthropic
    #   httpx
    #   openai
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.5
    # via requests
cryptography==46.0.5
    # via google-auth
distro==1.9.0
    # via
    #   anthropic
    #   openai
docstring-parser==0.17.0
    # via anthropic
google-api-core==2.30.0
    # via google-api-python-client
google-api-python-client==2.192.0
    # via -r requirements.in
google-auth==2.49.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.3.0
    # via
    #   -r requirements.in
    #   google-api-python-client
google-auth-oauthlib==1.3.0
    # via -r requirements.in
googleapis-common-protos==1.73.0
    # via google-api-core
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httplib2==0.31.2
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via
    #   anthropic
    #   openai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.13.0
    # via
    #   anthropic
    #   openai
oauthlib==3.3.1
    # via requests-oauthlib
openai==2.26.0
    # via -r requirements.in
proto-plus==1.27.1
    # via google-api-core
protobuf==6.33.5
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyaes==1.6.1
    # via telethon
pyasn1==0.6.2
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.23
    # via cffi
pydantic==2.12.5
    # via
    #   anthropic
    #   openai
pydantic-core==2.41.5
    # via pydantic
pyparsing==3.3.2
    # via httplib2
pyqrcode==1.2.1
    # via -r requirements.in
python-dotenv==1.2.1
    # via -r requirements.in
python-frontmatter==1.1.0
    # via -r requirements.in
pyyaml==6.0.3
    # via
    #   -r requirements.in
    #   python-frontmatter
requests==2.32.5
    # via
    #   -r requirements.in
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9.1
    # via
    #   google-auth
    #   telethon
sniffio==1.3.1
    # via
    #   anthropic
    #   openai
telethon==1.42.0
    # via -r requirements.in
tqdm==4.67.3
    # via openai
typing-extensions==4.15.0
    # via
    #   anthropic
    #   anyio
    #   openai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.6.3
    # via requests
